% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{query()}
\usage{
query(
  geoid = "36109",
  pollutant = 98,
  aggregation = 16,
  var = c("year", "vmt", "vehicles", "starts", "sourcehours"),
  sourcetype = NA,
  regclass = NA,
  fueltype = NA,
  roadtype = NA,
  base_url = default_base_url()
)
}
\arguments{
\item{base_url}{Base URL of the CAT Public API. Defaults to the
`moveslite.base_url` option, then the `MOVESLITE_BASE_URL` environment
variable, then `"https://api.cat-apps.com/"`. A trailing slash is optional,
so both `"https://x"` and `"https://x/"` work. Use this to point the client
at a replica, eg. `"https://connect.systems-apps.com/catplatform-public"`.}
}
\description{
Function to query catserver using CAT Public API.
}
\author{
Tim Fraser
}
