% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_default}
\alias{get_default}
\title{`get_default()`}
\usage{
get_default(.scenario = "granddata.d36109", .pollutant = 98, .by = "8.41")
}
\arguments{
\item{.scenario}{eg. granddata.d36109}

\item{.pollutant}{eg. 98}

\item{.by}{eg. 8.41 --> look at moveslite::choices_aggregation for values. Some are single values (by = 16), while others are combinations like 8.41 (by = 8 for sourcetype 41)}
}
\description{
Produces `default()` input data for a given CATSERVER query.
This data.frame is used to populate the original table and makes the benchmark scenario in the lineplot.
}
\author{
Tim
}
