% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{query()}
\usage{
query(
  geoid = "36109",
  pollutant = 98,
  aggregation = 16,
  var = c("year", "vmt", "vehicles", "starts", "sourcehours"),
  sourcetype = NA,
  regclass = NA,
  fueltype = NA,
  roadtype = NA
)
}
\description{
Function to query catserver using CAT Public API.
}
\author{
Tim Fraser
}
